Add support to hardware v1.3.1, always save network config
The hardware revision v1.3.1 comes with the 24AA02E48 eeprom connected
to the I2C0 bus to avoid slave address conflicts with the FM24CL16B
FERAM. The EEPROM comes with a unique EUI-48 address, no need to
manually set the MAC address for the first boot.

Save the network configuration to the FERAM after testing.